Adding namespace support

+++ setup/terraform/list-details.sh
@@ -1,46 +1,49 @@
 #!/bin/bash
-set -u
-set -e
+set -o errexit
+set -o nounset
 BASE_DIR=$(cd $(dirname $0); pwd -L)
+. $BASE_DIR/common.sh
 
-TF_JSON_FILE=.tf.json.$$
+if [ $# != 1 ]; then
+  echo "Syntax: $0 <namespace>"
+  show_namespaces
+  exit 1
+fi
+NAMESPACE=$1
+load_env $NAMESPACE
+
+TF_JSON_FILE=$BASE_DIR/.tf.json.$$
 trap "rm -f $TF_JSON_FILE" 0
 
-. $BASE_DIR/.env
-
-terraform show -json > $TF_JSON_FILE
+terraform show -json $NAMESPACE_DIR/terraform.state > $TF_JSON_FILE
 
 function web_instance() {
-  cat $TF_JSON_FILE | jq -r '.values[].resources[] | select(.type == "aws_instance" and .name == "web") | "\(.values.tags.Name) \(.values.public_dns) \(.values.public_ip)"'
+  cat $TF_JSON_FILE | jq -r '.values[]?.resources[]? | select(.type == "aws_instance" and .name == "web") | "\(.values.tags.Name) \(.values.public_dns) \(.values.public_ip) \(.values.private_ip)"'
 }
 
 function cluster_instances() {
-  cat $TF_JSON_FILE | jq -r '.values[].resources[] | select(.type == "aws_instance" and .name != "web") | "\(.values.tags.Name) \(.values.public_dns) \(.values.public_ip)"'
+  cat $TF_JSON_FILE | jq -r '.values[]?.resources[]? | select(.type == "aws_instance" and .name != "web") | "\(.values.tags.Name) \(.values.public_dns) \(.values.public_ip) \(.values.private_ip)"'
 }
 
-function web_key_file() {
-  echo $(cat $TF_JSON_FILE | jq -r '.values[].resources[] | select(.address == "aws_key_pair.workshop_web_key_pair").values.key_name').pem
-}
-
-function cluster_key_file() {
-  echo $(cat $TF_JSON_FILE | jq -r '.values[].resources[] | select(.address == "aws_key_pair.workshop_key_pair").values.key_name').pem
-}
-
-WEB_KEY_FILE=$(web_key_file)
-echo $WEB_KEY_FILE > .web.key.file.name
-echo "WEB SERVER Key file: $WEB_KEY_FILE"
-echo "WEB SERVER Key contents:"
-cat $WEB_KEY_FILE
+if [ -s "$TF_VAR_web_ssh_private_key" ]; then
+  echo "WEB SERVER Key file: $TF_VAR_web_ssh_private_key"
+  echo "WEB SERVER Key contents:"
+  cat $TF_VAR_web_ssh_private_key
+else
+  echo "WEB SERVER Key file is not available."
+fi
 echo ""
 
-KEY_FILE=$(cluster_key_file)
-echo $KEY_FILE > .key.file.name
-echo "Key file: $KEY_FILE"
-echo "Key contents:"
-cat $KEY_FILE
+if [ -s "$TF_VAR_ssh_private_key" ]; then
+  echo "Key file: $TF_VAR_ssh_private_key"
+  echo "Key contents:"
+  cat $TF_VAR_ssh_private_key
+else
+  echo "Key file is not available."
+fi
 echo ""
 
-echo "Web Server:       http://$(web_instance | awk '{print $NF}')"
+echo "Web Server:       http://$(web_instance | awk '{print $3}')"
 echo "Web Server admin: $TF_VAR_web_server_admin_email"
 echo ""
 
@@ -49,17 +52,17 @@ echo ""
 
 echo "WEB SERVER VM:"
 echo "=============="
-printf "%-40s %-55s %-15s\n" "Web Server Name" "Public DNS Name" "Public IP"
-web_instance | while read name public_dns public_ip; do
-  printf "%-40s %-55s %-15s\n" "$name" "$public_dns" "$public_ip"
-done | sed 's/\([^ ]*-\)\([0-9]*\)\( .*\)/\1\2\3 \2/' | sort -k4n | sed 's/ [0-9]*$//' | tee .instance.web
+printf "%-40s %-55s %-15s %-15s\n" "Web Server Name" "Public DNS Name" "Public IP" "Private IP"
+web_instance | while read name public_dns public_ip private_ip; do
+  printf "%-40s %-55s %-15s %-15s\n" "$name" "$public_dns" "$public_ip" "$private_ip"
+done | sed 's/\([^ ]*-\)\([0-9]*\)\( .*\)/\1\2\3 \2/' | sort -k4n | sed 's/ [0-9]*$//' | tee $WEB_INSTANCE_LIST_FILE
 echo ""
 
 echo "CLUSTER VMS:"
 echo "============"
-printf "%-40s %-55s %-15s\n" "Cluster Name" "Public DNS Name" "Public IP"
-cluster_instances | while read name public_dns public_ip; do
-  printf "%-40s %-55s %-15s\n" "$name" "$public_dns" "$public_ip"
-done | sed 's/\([^ ]*-\)\([0-9]*\)\( .*\)/\1\2\3 \2/' | sort -k4n | sed 's/ [0-9]*$//' | tee .instance.list
+printf "%-40s %-55s %-15s %-15s\n" "Cluster Name" "Public DNS Name" "Public IP" "Private IP"
+cluster_instances | while read name public_dns public_ip private_ip; do
+  printf "%-40s %-55s %-15s %-15s\n" "$name" "$public_dns" "$public_ip" "$private_ip"
+done | sed 's/\([^ ]*-\)\([0-9]*\)\( .*\)/\1\2\3 \2/' | sort -k4n | sed 's/ [0-9]*$//' | tee $INSTANCE_LIST_FILE
 
 echo ""
